Converts current console output to use verbose flag
