+++ packages/cli/package.json
@@ -27,6 +27,7 @@
     "is-valid-glob": "^1.0.0",
     "json-stable-stringify": "^1.0.1",
     "lodash": "^4.17.19",
+    "log-symbols": "^4.0.0",
     "p-map": "^4.0.0",
     "promise.hash.helper": "^1.0.7",
     "recast": "^0.20.4",
