+++ packages/cli/src/reporters/get-reporter.ts
@@ -1,9 +1,14 @@
-import { OutputFormat } from '@checkup/core';
+import { OutputFormat, RunFlags } from '@checkup/core';
+import { report as verboseConsoleReport } from './verbose-console-reporter';
 import { report as consoleReport } from './console-reporter';
 import { report as jsonReport } from './json-reporter';
 
-export function getReporter(outputFormat: OutputFormat) {
-  switch (outputFormat) {
+export function getReporter(runFlags: RunFlags) {
+  if (runFlags.verbose) {
+    return verboseConsoleReport;
+  }
+
+  switch (runFlags.format) {
     case OutputFormat.stdout: {
       return consoleReport;
     }
@@ -12,4 +17,6 @@ export function getReporter(outputFormat: OutputFormat) {
       return jsonReport;
     }
   }
+
+  throw new Error('No reporter found.');
 }
